pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plane-mcp-redux"
version = "0.1.0"
description = "A lean, lazy-loading MCP server for Plane project management"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "jax0m", email = "jax0m@users.noreply.github.com" }
]
keywords = ["mcp", "plane", "mcp-server", "fastmcp", "project-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=3.0.0",
    "plane-sdk>=0.2.8",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.27.0",
]

[project.scripts]
plane_rex = "plane_mcp.main:main"

[project.urls]
Homepage = "https://github.com/jax0m/plane-mcp-redux"
Repository = "https://github.com/jax0m/plane-mcp-redux"
Issues = "https://github.com/jax0m/plane-mcp-redux/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/plane_mcp"]

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501", "B008"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["plane_mcp"]

[tool.ty.rules]
invalid-assignment = "ignore"
invalid-argument-type = "ignore"
unknown-argument = "ignore"
missing-argument = "ignore"
unresolved-attribute = "ignore"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = false  # Allow unused ignores for SDK compatibility
disallow_untyped_defs = false  # SDK has incomplete types
disallow_incomplete_defs = false  # SDK has incomplete types
check_untyped_defs = false  # SDK has incomplete types
mypy_path = "src"
strict_optional = true

# Ignore SDK-related errors
[[tool.mypy.overrides]]
module = "plane.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/plane_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
